requirements.txt
azure-storage-blob==12.5.0
pycryptodome==3.9.8
cryptography==3.1.1
aiohttp==3.6.3
hvac==0.10.5
